+++ front/auth/app/common/auth-tracker.js
@@ -1,4 +1,4 @@
-import {getQueryParam} from 'common/utils/queryString';
+import {getQueryParam} from 'common/utils/querystring';
 import {track as mercuryTrack, trackActions, trackPageView as mercuryTrackPageView, setTrackContext}
 	from 'common/utils/track';
