+++ front/auth/app/facebook/facebook-login.js
@@ -1,9 +1,9 @@
-import AuthLogger from '../common/AuthLogger';
-import AuthTracker from '../common/AuthTracker';
-import AuthUtils from '../common/AuthUtils';
-import FacebookSDK from './FacebookSDK';
-import HttpCodes from '../common/HttpCodes';
-import UrlHelper from '../common/UrlHelper';
+import AuthLogger from '../common/auth-logger';
+import AuthTracker from '../common/auth-tracker';
+import AuthUtils from '../common/auth-utils';
+import FacebookSDK from './facebook-sdk';
+import HttpCodes from '../common/http-codes';
+import UrlHelper from '../common/url-helper';
 import {trackActions} from 'common/utils/track';
 
 /**
