+++ front/auth/app/tracking.js
@@ -1,7 +1,7 @@
 import AuthTracker from './common/auth-tracker';
 import UniversalAnalytics from 'common/modules/trackers/universal-analytics';
 import {trackActions} from 'common/utils/track';
-import {getQueryParam} from 'common/utils/queryString';
+import {getQueryParam} from 'common/utils/querystring';
 
 /**
  * @returns {void}
