+++ front/main/app/components/infobox-image-collection.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
 import MediaComponent from './media';
 import ViewportMixin from '../mixins/viewport';
-import Thumbnailer from 'common/modules/Thumbnailer';
+import Thumbnailer from 'common/modules/thumbnailer';
 
 export default MediaComponent.extend(
 	ViewportMixin,
