+++ front/main/app/components/smart-banner-android.js
@@ -1,5 +1,5 @@
 import Ember from 'ember';
-import Thumbnailer from 'common/modules/Thumbnailer';
+import Thumbnailer from 'common/modules/thumbnailer';
 import {track as mercuryTrack, trackActions} from 'common/utils/track';
 import {system, standalone} from 'common/utils/browser';
