+++ front/main/app/routes/application.js
@@ -3,7 +3,7 @@ import TrackClickMixin from '../mixins/track-click';
 import ArticleModel from '../models/wiki/article';
 import {activate as variantTestingActivate} from 'common/utils/variant-testing';
 import {normalizeToUnderscore} from 'common/utils/string';
-import Ads from 'common/modules/Ads';
+import Ads from 'common/modules/ads';
 import getLinkInfo from '../utils/article-link';
 
 const {
