+++ front/main/tests/index.html
@@ -30,6 +30,7 @@
 		};
 
 		M.prop('apiBase', '/api/mercury', true);
+		M.prop('environment', 'test', true);
 	</script>
 
 </head>
